requirements.txt
aiohttp==3.6.3
amqp==5.0.5
ansicon==1.89.0
arrow==0.17.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
billiard==3.6.3.0
blessed==1.17.12
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
discord==1.0.1
discord.py==1.5.1
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.7.0
django-heroku==0.3.1
django-kombu==0.9.4
django-picklefield==3.0.1
django-q==1.3.4
django-scheduler==0.9.3
django-timezone-field==4.1.1
djangorestframework==3.12.2
drf-nested-serializer==0.1.0
export==0.1.2
gunicorn==20.0.4
icalendar==4.0.7
idna==2.10
jinxed==1.0.1
kombu==5.0.2
lxml==4.6.2
MarkupSafe==1.1.1
multidict==4.7.6
npm==0.1.1
optional-django==0.1.0
Pillow==8.0.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
